% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{remove_na_rows}
\alias{remove_na_rows}
\title{Remove NA values from a dataframe}
\usage{
remove_na_rows(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
Clean dataframe
}
\description{
Remove NA values from a dataframe
}
